+++ CHANGELOG.md
@@ -309,6 +309,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `\OxidEsales\EshopCommunity\Application\Controller\Admin\AdminDetailsController::_getPlainEditor()`, use TextEditorHandler::renderPlainTextEditor()
 - `\OxidEsales\EshopCommunity\Application\Controller\Admin\AdminDetailsController::_generateTextEditor()`, use generateTextEditor()
 - `\OxidEsales\EshopCommunity\Application\Model\Order::getFormattedeliveryCost()`, use getFormattedDeliveryCost()
+- `\OxidEsales\EshopCommunity\Application\Model\Delivery::_isForArticle`, use isDeliveryRuleFitByArticle()
 
 ### Fixed
 - Fix not working actions and promotions [#0005526](https://bugs.oxid-esales.com/view.php?id=5526)